add custom icon metadata to taxonomy terms
each tag now has a yaml front-matter file with an iconClassList
parameter that associates each tag in the taxonomy with a font-awesome
icon. the varaible is the CSS classlist necessary to render that tag's
associated icon.

this also cleans up the tagging in a few posts to match the clean tags
list